# replace github.com/minio/blake2b-simd with golang.org/x/crypto/blake2b