… vendor/github.com/minio/blake2b-simd/.gitignore
… vendor/github.com/minio/blake2b-simd/.travis.yml
… vendor/github.com/minio/blake2b-simd/LICENSE
… vendor/github.com/minio/blake2b-simd/README.md
… vendor/github.com/minio/blake2b-simd/appveyor.yml
… vendor/github.com/minio/blake2b-simd/benchmarks_test.go
… vendor/github.com/minio/blake2b-simd/blake2b.go
… vendor/github.com/minio/blake2b-simd/blake2b_test.go
… vendor/github.com/minio/blake2b-simd/compressAvx2_amd64.go
… vendor/github.com/minio/blake2b-simd/compressAvx2_amd64.s
… vendor/github.com/minio/blake2b-simd/compressAvx_amd64.go
… vendor/github.com/minio/blake2b-simd/compressAvx_amd64.s
… vendor/github.com/minio/blake2b-simd/compressSse_amd64.go
… vendor/github.com/minio/blake2b-simd/compressSse_amd64.s
… vendor/github.com/minio/blake2b-simd/compress_amd64.go
… vendor/github.com/minio/blake2b-simd/compress_generic.go
… vendor/github.com/minio/blake2b-simd/compress_noasm.go
… vendor/github.com/minio/blake2b-simd/cpuid.go
… vendor/github.com/minio/blake2b-simd/cpuid_386.s
… vendor/github.com/minio/blake2b-simd/cpuid_amd64.s
… vendor/golang.org/x/crypto/blake2b/blake2b.go
… vendor/golang.org/x/crypto/blake2b/blake2bAVX2_amd64.go
… vendor/golang.org/x/crypto/blake2b/blake2bAVX2_amd64.s
… vendor/golang.org/x/crypto/blake2b/blake2b_amd64.go
… vendor/golang.org/x/crypto/blake2b/blake2b_amd64.s
… vendor/golang.org/x/crypto/blake2b/blake2b_generic.go
… vendor/golang.org/x/crypto/blake2b/blake2b_ref.go
… vendor/golang.org/x/crypto/blake2b/blake2b_test.go
… vendor/golang.org/x/crypto/blake2b/blake2x.go
… vendor/golang.org/x/crypto/blake2b/register.go