closes #510 
results of benchmark before and after change of the lib:
```
benchmark                                               old ns/op     new ns/op     delta
BenchmarkSumBlake2b256/512-bytes-4                      1123          917           -18.34%
BenchmarkSumBlake2b256/4096-bytes-4                     7056          6565          -6.96%
BenchmarkSumBlake2b256/131072-bytes-4                   216805        207368        -4.35%
BenchmarkSumBlake2b512/512-bytes-4                      1153          929           -19.43%
BenchmarkSumBlake2b512/4096-bytes-4                     7167          6607          -7.81%
BenchmarkSumBlake2b512/131072-bytes-4                   216676        207697        -4.14%
BenchmarkSumBlake2b256er_WithoutKey/512-bytes-4         1105          960           -13.12%
BenchmarkSumBlake2b256er_WithoutKey/4096-bytes-4        7105          6676          -6.04%
BenchmarkSumBlake2b256er_WithoutKey/131072-bytes-4      217568        207869        -4.46%
BenchmarkSumBlake2b256er_WithoutKey/1048576-bytes-4     217467        207658        -4.51%
BenchmarkSumBlake2b256er_WithKey/512-bytes-4            1394          1195          -14.28%
BenchmarkSumBlake2b256er_WithKey/4096-bytes-4           7768          6876          -11.48%
BenchmarkSumBlake2b256er_WithKey/131072-bytes-4         218767        208060        -4.89%
BenchmarkSumBlake2b256er_WithKey/1048576-bytes-4        1752674       1660718       -5.25%
BenchmarkSumBlake2b512er_WithoutKey/512-bytes-4         1145          990           -13.54%
BenchmarkSumBlake2b512er_WithoutKey/4096-bytes-4        7080          6691          -5.49%
BenchmarkSumBlake2b512er_WithoutKey/131072-bytes-4      216759        207399        -4.32%
BenchmarkSumBlake2b512er_WithoutKey/1048576-bytes-4     1730830       1654038       -4.44%
BenchmarkSumBlake2b512er_WithKey/512-bytes-4            1436          1211          -15.67%
BenchmarkSumBlake2b512er_WithKey/4096-bytes-4           7371          6883          -6.62%
BenchmarkSumBlake2b512er_WithKey/131072-bytes-4         216894        208178        -4.02%
BenchmarkSumBlake2b512er_WithKey/1048576-bytes-4        1738029       1659043       -4.54%
```