@GlenDC 2018-02-22
Does the hardware you ran the benchmarks on support AVX2, @zaibon?