@zaibon 2018-02-22
> Does the hardware you ran the benchmarks on support AVX2, @zaibon?

Yes it does

I re-pushed a version after updated my dep tool, but it doesn't seems to have change the layout.