@GlenDC 2018-03-07
> Seems there is a problem because of https://github.com/zero-os/0-stor/pull/518/files#diff-bd247e83efc3c45ae9e8c47233249f18R270
Seesm the full crypto package is locked to a specific revision. I tried to change that, but then it lots of other stuff, which make me think that this specific revision is needed for grpc or another dependencies.

Not sure what you're talking about here, all I had to change was that you need to use the correct constructor provided by google's lib, rather than the lib we were using.